src/requirements.txt
django==4.1
google-cloud-secret-manager==2.12.0
django-environ==0.9.0
django-ckeditor==6.5.0